🐛 Unpack GameParams in-app from the game pkg store and add manual game-path setup. (#371)
- Root cause: modern WoWS keeps GameParams.data inside bin/<build> idx/pkg
  archives; the old lookup only checked loose legacy paths, so every fresh
  player install failed armor/ballistics loads even with a correct game
  root (dev machines worked only via the extract script's LOCALAPPDATA
  cache). The bottom-left process indicator kept working because it reads
  the running exe directly.
- gameparams: mount the vendored wowsunpack VFS (new workspace dep,
  vfs-mmap only), read content/GameParams.data, decode the pickle
  (modern {"": ...} wrapper, object __dict__ entries, list/tuple roots),
  extract the per-ship slice, and cache under AppData keyed to the current
  bin build so slices drop when the game updates. Verified end-to-end
  against a real Steam install (PASB510_Ohio, 255KB slice).
- game_detect: new pick_game_folder command (rfd native folder picker);
  set_game_path now fills realm and returns an actionable error.
- get_game_process synthesizes a matched install from a running exe that
  no detected install claims, unifying the process indicator with the
  data loaders.
- webui: GamePathSetupModal (detected install list + native picker +
  running-game shortcut + re-detect), auto-popped at startup whenever the
  detect pass ends without an active install, reachable from Settings and
  the ship-detail armor-error banner (which auto-retries when the path
  changes). ShipsView falls back to the running process's install;
  remembered manual paths are re-validated on rescan so they survive
  restarts instead of re-triggering the prompt every launch.
- i18n: common.retry + common.gamePath.* + settings.gamePath* across all
  9 locales; mock backend stub for pick_game_folder.
- clippy scoped to the app crates with --no-deps in justfile + CI (the
  vendored wowsunpack is now in the dep graph and keeps upstream lint
  style); fixed three pre-existing all-targets clippy lints so just fmt
  passes again.

+++ packages/app/tauri/src/commands/overlay_detect.rs
@@ -708,7 +708,7 @@ mod tests {
         let det = detect_roster(&img, w, h, (12, 12)).expect("header alone must anchor");
         assert_eq!(det.row_centers.len(), 24, "12 + 12 centers");
         // The 6 drawn rows must coincide with grid rows 0..6.
-        let pitch = (h as f32 * 0.028 * 0.92);
+        let pitch = h as f32 * 0.028 * 0.92;
         let first = (h as f32 * 0.22 + h as f32 * 0.028) + pitch * 0.5;
         for (k, &c) in det.row_centers.iter().enumerate().take(6) {
             let truth = first + pitch * k as f32;
@@ -848,7 +848,7 @@ mod tests {
             width: 1200,
             height: 250,
         };
-        let rows: Vec<i32> = (0..5).map(|i| 300 + 40 * i as i32 + 20).collect();
+        let rows: Vec<i32> = (0..5).map(|i| 300 + 40 * i + 20).collect();
         let (overlay, anchor) = build_anchor(&game, &roster, rows.clone(), 0.5, false);
         assert!(!anchor.table_detected);
         // The overlay window covers ONLY the inflated table area — wider on
